LayoutTests/js/excessive-comma-usage-expected.txt

Test that we can handle excessively large initializer lists

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS new Function(initializerTestString)() is true
PASS new Function(declarationTestString)() is true
PASS new Function(commaExpressionTestString)() is true
PASS successfullyParsed is true

TEST COMPLETE
